backend/requirements.txt
flask
flask-cors
duckdb
google-generativeai
openai
fastapi
uvicorn 